+++ packages/test/CHANGELOG.md
@@ -34,6 +34,7 @@
 - Updated dependencies [70e313e]
   - @confect/core@10.0.0-next.0
   - @confect/server@10.0.0-next.0
+
 ## 9.2.4
 
 ## 9.2.3